requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.1.5
django_debug_toolbar==3.8.1
filelock==3.9.0
mysqlclient==2.1.1
Pillow==9.4.0
platformdirs==2.6.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.17.1
whitenoise==6.3.0
